+++ test/src/features/transactions/transactions_list/view/transactions_list_page_test.dart
@@ -8,7 +8,7 @@ import 'package:vinttem_app/src/features/transactions/transaction.dart'
     hide Transaction, TransactionCategory, TransactionType, TransactionUser;
 import 'package:vinttem_repository/vinttem_repository.dart';
 
-import '../../../helpers/helpers.dart';
+import '../../../../../helpers/helpers.dart';
 
 class MockVinttemRepository extends Mock implements VinttemRepository {}
